+++ src/OpenTelemetry.AutoInstrumentation/Constants.cs
@@ -20,7 +20,7 @@ internal static class Constants
 {
     public static class Tracer
     {
-        public const string Version = "0.5.1-beta.2";
+        public const string Version = "0.5.1-beta.3";
         public const string AutoInstrumentationVersionName = "telemetry.auto.version";
     }
 
@@ -45,7 +45,6 @@ internal static class Constants
             public const string Otlp = "otlp";
             public const string Prometheus = "prometheus";
             public const string Zipkin = "zipkin";
-            public const string Jaeger = "jaeger";
         }
 
         public static class Propagators