+++ test/OpenTelemetry.AutoInstrumentation.Tests/Configuration/SettingsTests.cs
@@ -126,7 +126,6 @@ public class SettingsTests : IDisposable
 
     [Theory]
     [InlineData("none", TracesExporter.None)]
-    [InlineData("jaeger", TracesExporter.Jaeger)]
     [InlineData("otlp", TracesExporter.Otlp)]
     [InlineData("zipkin", TracesExporter.Zipkin)]
     internal void TracesExporter_SupportedValues(string tracesExporter, TracesExporter expectedTracesExporter)
@@ -195,6 +194,7 @@ public class SettingsTests : IDisposable
     [InlineData(nameof(TracerInstrumentation.MassTransit), TracerInstrumentation.MassTransit)]
 #endif
     [InlineData(nameof(TracerInstrumentation.NServiceBus), TracerInstrumentation.NServiceBus)]
+    [InlineData(nameof(TracerInstrumentation.Elasticsearch), TracerInstrumentation.Elasticsearch)]
     internal void TracerSettings_Instrumentations_SupportedValues(string tracerInstrumentation, TracerInstrumentation expectedTracerInstrumentation)
     {
         Environment.SetEnvironmentVariable(ConfigurationKeys.Traces.Instrumentations, tracerInstrumentation);